Renamed some files for consistency and clarity. Updated README.md
