+++ custom-platform/group-manager.js
